Added to room entity support for JPA properties and some ui representation
TODO -add final support for rooms
     -add support for chekins(booking)
     -add support for admin system

+++ src/main/java/com/example/hotel_jpa/repositories/IRoomRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface IRoomRepository extends JpaRepository<Room, Integer> {
+public interface IRoomRepository extends JpaRepository<Room, Long> {
 
     @Query("select r from Room r where r.number = :number")
     Room findByNumber(@Param("number") Integer number);
